Flask_work_folder/jiomart_clone/requirements.txt
Flask
mysql-connector-python
PyJWT
Werkzeug